aikshetraofficemanagementsystemqlpubjalphaproject/backend/requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Database - MongoDB
djongo==1.3.6
pymongo==4.6.0
dnspython==2.4.2
sqlparse==0.2.4

# Authentication and JWT
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Real-time support
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# File uploads and media
Pillow>=10.1.0

# Development tools
django-debug-toolbar==4.2.0

# API documentation
drf-spectacular==0.26.5

# Utilities
python-dateutil==2.8.2

python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
